+++ lib/api/routes/index.ts
@@ -4,3 +4,4 @@ export * from "./info";
 export * from "./fetch";
 export * from "./produce";
 export * from "./manage";
+export * from "./hook";