+++ lib/db/models/index.ts
@@ -2,3 +2,4 @@ export * from "./KeyIndexModel";
 export * from "./TopicConfigModel";
 export * from "./TopicMetadata";
 export * from "./LockModel";
+export * from "./HookModel";